requirements.txt
streamlit>=1.29.0
pdfplumber>=0.10.0
pandas>=2.0.0
python-dateutil>=2.8.0
regex>=2023.0.0
sentence-transformers>=4.1.0
plotly>=5.17.0
numpy>=1.24.0
PyPDF2>=3.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
huggingface-hub>=0.20.0
transformers>=4.36.0
torch>=2.0.0
opencv-python-headless>=4.8.0
money-parser==0.0.5
datefinder==0.7.3
pdf2image==1.16.3 